refactor: rename player-setup.ts to plyr-setup.ts
