+++ rollup.config.js
@@ -2,7 +2,7 @@ import typescript from "@rollup/plugin-typescript";
 import { nodeResolve } from "@rollup/plugin-node-resolve";
 import commonjs from "@rollup/plugin-commonjs";
 import css from "rollup-plugin-import-css";
-import json from '@rollup/plugin-json';
+import json from "@rollup/plugin-json";
 
 const isProd = process.env.BUILD === "production";
